+++ src/Stratis.Bitcoin/P2P/PeerConnector.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Threading;
@@ -11,7 +10,6 @@ using Stratis.Bitcoin.Configuration.Settings;
 using Stratis.Bitcoin.Connection;
 using Stratis.Bitcoin.P2P.Peer;
 using Stratis.Bitcoin.Utilities;
-using Stratis.Bitcoin.Utilities.Extensions;
 
 namespace Stratis.Bitcoin.P2P
 {
@@ -56,7 +54,7 @@ namespace Stratis.Bitcoin.P2P
         /// <summary>
         /// Collection of connected peers that is managed by the <see cref="ConnectionManager"/>.
         /// </summary>
-        private IReadOnlyNetworkPeerCollection connectedPeers;
+        protected IConnectionManager connectionManager;
 
         /// <inheritdoc/>
         public NetworkPeerCollection ConnectorPeers { get; private set; }
@@ -137,7 +135,7 @@ namespace Stratis.Bitcoin.P2P
         /// <inheritdoc/>
         public void Initialize(IConnectionManager connectionManager)
         {
-            this.connectedPeers = connectionManager.ConnectedPeers;
+            this.connectionManager = connectionManager;
 
             this.CurrentParameters = connectionManager.Parameters.Clone();
 
@@ -194,7 +192,7 @@ namespace Stratis.Bitcoin.P2P
         /// <param name="ipEndpoint">The endpoint to check.</param>
         internal bool IsPeerConnected(IPEndPoint ipEndpoint)
         {
-            return this.connectedPeers.FindByEndpoint(ipEndpoint) != null;
+            return this.connectionManager.ConnectedPeers.FindByEndpoint(ipEndpoint) != null;
         }
 
         /// <inheritdoc/>
@@ -225,13 +223,6 @@ namespace Stratis.Bitcoin.P2P
             if (this.selfEndpointTracker.IsSelf(peerAddress.Endpoint))
             {
                 this.logger.LogTrace("{0} is self. Therefore not connecting.", peerAddress.Endpoint);
-                return;
-            }
-
-            // Connect if local, ip range filtering disabled or ip range filtering enabled and peer in a different group.
-            if (peerAddress.Endpoint.Address.IsRoutable(false) && this.ConnectionSettings.IpRangeFiltering && this.PeerIsPartOfExistingGroup(peerAddress))
-            {
-                this.logger.LogTrace("(-)[RANGE_FILTERED]");
                 return;
             }
 
@@ -279,30 +270,6 @@ namespace Stratis.Bitcoin.P2P
                 peerAddress.SetHandshakeAttempted(this.dateTimeProvider.GetUtcNow());
                 peer?.Disconnect("Error while connecting", exception);
             }
-        }
-
-        private bool PeerIsPartOfExistingGroup(PeerAddress peerAddress)
-        {
-            if (this.connectedPeers == null)
-            {
-                this.logger.LogTrace("(-)[NO_CONNECTED_PEERS]:false");
-                return false;
-            }
-
-            byte[] peerAddressGroup = peerAddress.Endpoint.MapToIpv6().Address.GetGroup();
-
-            foreach (INetworkPeer endPoint in this.connectedPeers)
-            {
-                byte[] endPointGroup = endPoint.PeerEndPoint.MapToIpv6().Address.GetGroup();
-                if (endPointGroup.SequenceEqual(peerAddressGroup))
-                {
-                    this.logger.LogTrace("(-)[SAME_GROUP]:true");
-                    return true;
-                }
-            }
-
-            this.logger.LogTrace("(-):false");
-            return false;
         }
 
         /// <summary>
