+++ src/Stratis.Bitcoin/P2P/PeerConnectorDiscovery.cs
@@ -128,7 +128,40 @@ namespace Stratis.Bitcoin.P2P
                 await Task.Delay(2000, this.nodeLifetime.ApplicationStopping).ConfigureAwait(false);
             }
             else
+            {
+                // Connect if local, ip range filtering disabled or ip range filtering enabled and peer in a different group.
+                if (peer.Endpoint.Address.IsRoutable(false) && this.ConnectionSettings.IpRangeFiltering && this.PeerIsPartOfExistingGroup(peer))
+                {
+                    this.logger.LogTrace("(-)[RANGE_FILTERED]");
+                    return;
+                }
+
                 await this.ConnectAsync(peer).ConfigureAwait(false);
+            }
+        }
+
+        private bool PeerIsPartOfExistingGroup(PeerAddress peerAddress)
+        {
+            if (this.connectionManager.ConnectedPeers == null)
+            {
+                this.logger.LogTrace("(-)[NO_CONNECTED_PEERS]:false");
+                return false;
+            }
+
+            byte[] peerAddressGroup = peerAddress.Endpoint.MapToIpv6().Address.GetGroup();
+
+            foreach (INetworkPeer endPoint in this.connectionManager.ConnectedPeers.ToList())
+            {
+                byte[] endPointGroup = endPoint.PeerEndPoint.MapToIpv6().Address.GetGroup();
+                if (endPointGroup.SequenceEqual(peerAddressGroup))
+                {
+                    this.logger.LogTrace("(-)[SAME_GROUP]:true");
+                    return true;
+                }
+            }
+
+            this.logger.LogTrace("(-):false");
+            return false;
         }
     }
 }
